pyproject.toml
[project]
name = "llm2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "html2text>=2024.2.26",
    "httpx==0.27.2",
    "langchain-chroma>=0.2.0",
    "langchain-community==0.2.5",
    "langchain-google-community<2",
    "langchain-google-genai>=1.0.10",
    "langchain-huggingface>=0.0.3",
    "langserve[all]>=0.2.3",
    "python-dotenv>=1.0.1",
]
